+++ src/api/index.ts
@@ -1,3 +1,4 @@
 export * from './getAllArchiveVos';
 export * from './getArchiveRootFolderVo';
 export * from './getFolderVo';
+export * from './getRecordVo';