+++ src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './folderVoToFolder';
 export * from './formatTimestampAsUtc';
 export * from './generateEnumTypeguard';
 export * from './makePermanentApiCall';
+export * from './recordVoToRecord';