+++ src/utils/recordVoToRecord.ts
@@ -25,10 +25,11 @@ export const recordVoToRecord = (recordVo: RecordVo): Record => {
   const displayDate = new Date(formatTimestampAsUtc(recordVo.displayDT));
   const type = recordTypeToRecordType(recordVo.type);
   const status = recordStatusToStatus(recordVo.status);
-  const files = (recordVo.fileVos ?? []).map((fileVo) => fileVoToFile(fileVo));
+  const files = (recordVo.FileVOs ?? []).map((fileVo) => fileVoToFile(fileVo));
   return {
     id: recordVo.recordId,
-    name: recordVo.displayName,
+    displayName: recordVo.displayName,
+    fileName: recordVo.downloadName,
     type,
     status,
     files,